refactor(tests): Split updateSubtaskById tests
