+++ tests/unit/task-manager.test.js
@@ -285,750 +285,4 @@ describe('Task Manager Module', () => {
 	// }); // REMOVE THIS BLOCK END
 });
 
-// Mock implementation of updateSubtaskById for testing
-const testUpdateSubtaskById = async (
-	tasksPath,
-	subtaskId,
-	prompt,
-	useResearch = false
-) => {
-	try {
-		// Parse parent and subtask IDs
-		if (
-			!subtaskId ||
-			typeof subtaskId !== 'string' ||
-			!subtaskId.includes('.')
-		) {
-			throw new Error(`Invalid subtask ID format: ${subtaskId}`);
-		}
-
-		const [parentIdStr, subtaskIdStr] = subtaskId.split('.');
-		const parentId = parseInt(parentIdStr, 10);
-		const subtaskIdNum = parseInt(subtaskIdStr, 10);
-
-		if (
-			isNaN(parentId) ||
-			parentId <= 0 ||
-			isNaN(subtaskIdNum) ||
-			subtaskIdNum <= 0
-		) {
-			throw new Error(`Invalid subtask ID format: ${subtaskId}`);
-		}
-
-		// Validate prompt
-		if (!prompt || typeof prompt !== 'string' || prompt.trim() === '') {
-			throw new Error('Prompt cannot be empty');
-		}
-
-		// Check if tasks file exists
-		if (!mockExistsSync(tasksPath)) {
-			throw new Error(`Tasks file not found at path: ${tasksPath}`);
-		}
-
-		// Read the tasks file
-		const data = mockReadJSON(tasksPath);
-		if (!data || !data.tasks) {
-			throw new Error(`No valid tasks found in ${tasksPath}`);
-		}
-
-		// Find the parent task
-		const parentTask = data.tasks.find((t) => t.id === parentId);
-		if (!parentTask) {
-			throw new Error(`Parent task with ID ${parentId} not found`);
-		}
-
-		// Find the subtask
-		if (!parentTask.subtasks || !Array.isArray(parentTask.subtasks)) {
-			throw new Error(`Parent task ${parentId} has no subtasks`);
-		}
-
-		const subtask = parentTask.subtasks.find((st) => st.id === subtaskIdNum);
-		if (!subtask) {
-			throw new Error(`Subtask with ID ${subtaskId} not found`);
-		}
-
-		// Check if subtask is already completed
-		if (subtask.status === 'done' || subtask.status === 'completed') {
-			return null;
-		}
-
-		// Generate additional information
-		let additionalInformation;
-		if (useResearch) {
-			const result = await mockChatCompletionsCreate();
-			additionalInformation = result.choices[0].message.content;
-		} else {
-			const mockStream = {
-				[Symbol.asyncIterator]: jest.fn().mockImplementation(() => {
-					return {
-						next: jest
-							.fn()
-							.mockResolvedValueOnce({
-								done: false,
-								value: {
-									type: 'content_block_delta',
-									delta: { text: 'Additional information about' }
-								}
-							})
-							.mockResolvedValueOnce({
-								done: false,
-								value: {
-									type: 'content_block_delta',
-									delta: { text: ' the subtask implementation.' }
-								}
-							})
-							.mockResolvedValueOnce({ done: true })
-					};
-				})
-			};
-
-			const stream = await mockCreate();
-			additionalInformation =
-				'Additional information about the subtask implementation.';
-		}
-
-		// Create timestamp
-		const timestamp = new Date().toISOString();
-
-		// Format the additional information with timestamp
-		const formattedInformation = `\n\n<info added on ${timestamp}>\n${additionalInformation}\n</info added on ${timestamp}>`;
-
-		// Append to subtask details
-		if (subtask.details) {
-			subtask.details += formattedInformation;
-		} else {
-			subtask.details = formattedInformation;
-		}
-
-		// Update description with update marker for shorter updates
-		if (subtask.description && additionalInformation.length < 200) {
-			subtask.description += ` [Updated: ${new Date().toLocaleDateString()}]`;
-		}
-
-		// Write the updated tasks to the file
-		mockWriteJSON(tasksPath, data);
-
-		// Generate individual task files
-		await mockGenerateTaskFiles(tasksPath, path.dirname(tasksPath));
-
-		return subtask;
-	} catch (error) {
-		mockLog('error', `Error updating subtask: ${error.message}`);
-		return null;
-	}
-};
-
-describe.skip('updateSubtaskById function', () => {
-	let mockConsoleLog;
-	let mockConsoleError;
-	let mockProcess;
-
-	beforeEach(() => {
-		// Reset all mocks
-		jest.clearAllMocks();
-
-		// Set up default mock values
-		mockExistsSync.mockReturnValue(true);
-		mockWriteJSON.mockImplementation(() => {});
-		mockGenerateTaskFiles.mockResolvedValue(undefined);
-
-		// Create a deep copy of sample tasks for tests - use imported ES module instead of require
-		const sampleTasksDeepCopy = JSON.parse(JSON.stringify(sampleTasks));
-
-		// Ensure the sample tasks has a task with subtasks for testing
-		// Task 3 should have subtasks
-		if (sampleTasksDeepCopy.tasks && sampleTasksDeepCopy.tasks.length > 2) {
-			const task3 = sampleTasksDeepCopy.tasks.find((t) => t.id === 3);
-			if (task3 && (!task3.subtasks || task3.subtasks.length === 0)) {
-				task3.subtasks = [
-					{
-						id: 1,
-						title: 'Create Header Component',
-						description: 'Create a reusable header component',
-						status: 'pending'
-					},
-					{
-						id: 2,
-						title: 'Create Footer Component',
-						description: 'Create a reusable footer component',
-						status: 'pending'
-					}
-				];
-			}
-		}
-
-		mockReadJSON.mockReturnValue(sampleTasksDeepCopy);
-
-		// Mock console and process.exit
-		mockConsoleLog = jest.spyOn(console, 'log').mockImplementation(() => {});
-		mockConsoleError = jest
-			.spyOn(console, 'error')
-			.mockImplementation(() => {});
-		mockProcess = jest.spyOn(process, 'exit').mockImplementation(() => {});
-	});
-
-	afterEach(() => {
-		// Restore console and process.exit
-		mockConsoleLog.mockRestore();
-		mockConsoleError.mockRestore();
-		mockProcess.mockRestore();
-	});
-
-	test('should update a subtask successfully', async () => {
-		// Mock streaming for successful response
-		const mockStream = {
-			[Symbol.asyncIterator]: jest.fn().mockImplementation(() => {
-				return {
-					next: jest
-						.fn()
-						.mockResolvedValueOnce({
-							done: false,
-							value: {
-								type: 'content_block_delta',
-								delta: {
-									text: 'Additional information about the subtask implementation.'
-								}
-							}
-						})
-						.mockResolvedValueOnce({ done: true })
-				};
-			})
-		};
-
-		mockCreate.mockResolvedValue(mockStream);
-
-		// Call the function
-		const result = await testUpdateSubtaskById(
-			'test-tasks.json',
-			'3.1',
-			'Add details about API endpoints'
-		);
-
-		// Verify the subtask was updated
-		expect(result).toBeDefined();
-		expect(result.details).toContain('<info added on');
-		expect(result.details).toContain(
-			'Additional information about the subtask implementation'
-		);
-		expect(result.details).toContain('</info added on');
-
-		// Verify the correct functions were called
-		expect(mockReadJSON).toHaveBeenCalledWith('test-tasks.json');
-		expect(mockCreate).toHaveBeenCalled();
-		expect(mockWriteJSON).toHaveBeenCalled();
-		expect(mockGenerateTaskFiles).toHaveBeenCalled();
-
-		// Verify the subtask was updated in the tasks data
-		const tasksData = mockWriteJSON.mock.calls[0][1];
-		const parentTask = tasksData.tasks.find((task) => task.id === 3);
-		const updatedSubtask = parentTask.subtasks.find((st) => st.id === 1);
-		expect(updatedSubtask.details).toContain(
-			'Additional information about the subtask implementation'
-		);
-	});
-
-	test('should return null when subtask is already completed', async () => {
-		// Modify the sample data to have a completed subtask
-		const tasksData = mockReadJSON();
-		const task = tasksData.tasks.find((t) => t.id === 3);
-		if (task && task.subtasks && task.subtasks.length > 0) {
-			// Mark the first subtask as completed
-			task.subtasks[0].status = 'done';
-			mockReadJSON.mockReturnValue(tasksData);
-		}
-
-		// Call the function with a completed subtask
-		const result = await testUpdateSubtaskById(
-			'test-tasks.json',
-			'3.1',
-			'Update completed subtask'
-		);
-
-		// Verify the result is null
-		expect(result).toBeNull();
-
-		// Verify the correct functions were called
-		expect(mockReadJSON).toHaveBeenCalledWith('test-tasks.json');
-		expect(mockCreate).not.toHaveBeenCalled();
-		expect(mockWriteJSON).not.toHaveBeenCalled();
-		expect(mockGenerateTaskFiles).not.toHaveBeenCalled();
-	});
-
-	test('should handle subtask not found error', async () => {
-		// Call the function with a non-existent subtask
-		const result = await testUpdateSubtaskById(
-			'test-tasks.json',
-			'3.999',
-			'Update non-existent subtask'
-		);
-
-		// Verify the result is null
-		expect(result).toBeNull();
-
-		// Verify the error was logged
-		expect(mockLog).toHaveBeenCalledWith(
-			'error',
-			expect.stringContaining('Subtask with ID 3.999 not found')
-		);
-
-		// Verify the correct functions were called
-		expect(mockReadJSON).toHaveBeenCalledWith('test-tasks.json');
-		expect(mockCreate).not.toHaveBeenCalled();
-		expect(mockWriteJSON).not.toHaveBeenCalled();
-		expect(mockGenerateTaskFiles).not.toHaveBeenCalled();
-	});
-
-	test('should handle invalid subtask ID format', async () => {
-		// Call the function with an invalid subtask ID
-		const result = await testUpdateSubtaskById(
-			'test-tasks.json',
-			'invalid-id',
-			'Update subtask with invalid ID'
-		);
-
-		// Verify the result is null
-		expect(result).toBeNull();
-
-		// Verify the error was logged
-		expect(mockLog).toHaveBeenCalledWith(
-			'error',
-			expect.stringContaining('Invalid subtask ID format')
-		);
-
-		// Verify the correct functions were called
-		expect(mockReadJSON).toHaveBeenCalledWith('test-tasks.json');
-		expect(mockCreate).not.toHaveBeenCalled();
-		expect(mockWriteJSON).not.toHaveBeenCalled();
-		expect(mockGenerateTaskFiles).not.toHaveBeenCalled();
-	});
-
-	test('should handle missing tasks file', async () => {
-		// Mock file not existing
-		mockExistsSync.mockReturnValue(false);
-
-		// Call the function
-		const result = await testUpdateSubtaskById(
-			'missing-tasks.json',
-			'3.1',
-			'Update subtask'
-		);
-
-		// Verify the result is null
-		expect(result).toBeNull();
-
-		// Verify the error was logged
-		expect(mockLog).toHaveBeenCalledWith(
-			'error',
-			expect.stringContaining('Tasks file not found')
-		);
-
-		// Verify the correct functions were called
-		expect(mockReadJSON).not.toHaveBeenCalled();
-		expect(mockCreate).not.toHaveBeenCalled();
-		expect(mockWriteJSON).not.toHaveBeenCalled();
-		expect(mockGenerateTaskFiles).not.toHaveBeenCalled();
-	});
-
-	test('should handle empty prompt', async () => {
-		// Call the function with an empty prompt
-		const result = await testUpdateSubtaskById('test-tasks.json', '3.1', '');
-
-		// Verify the result is null
-		expect(result).toBeNull();
-
-		// Verify the error was logged
-		expect(mockLog).toHaveBeenCalledWith(
-			'error',
-			expect.stringContaining('Prompt cannot be empty')
-		);
-
-		// Verify the correct functions were called
-		expect(mockReadJSON).toHaveBeenCalledWith('test-tasks.json');
-		expect(mockCreate).not.toHaveBeenCalled();
-		expect(mockWriteJSON).not.toHaveBeenCalled();
-		expect(mockGenerateTaskFiles).not.toHaveBeenCalled();
-	});
-
-	test('should use Perplexity AI when research flag is true', async () => {
-		// Mock Perplexity API response
-		const mockPerplexityResponse = {
-			choices: [
-				{
-					message: {
-						content:
-							'Research-backed information about the subtask implementation.'
-					}
-				}
-			]
-		};
-
-		mockChatCompletionsCreate.mockResolvedValue(mockPerplexityResponse);
-
-		// Set the Perplexity API key in environment
-		process.env.PERPLEXITY_API_KEY = 'dummy-key';
-
-		// Call the function with research flag
-		const result = await testUpdateSubtaskById(
-			'test-tasks.json',
-			'3.1',
-			'Add research-backed details',
-			true
-		);
-
-		// Verify the subtask was updated with research-backed information
-		expect(result).toBeDefined();
-		expect(result.details).toContain('<info added on');
-		expect(result.details).toContain(
-			'Research-backed information about the subtask implementation'
-		);
-		expect(result.details).toContain('</info added on');
-
-		// Verify the Perplexity API was called
-		expect(mockChatCompletionsCreate).toHaveBeenCalled();
-		expect(mockCreate).not.toHaveBeenCalled(); // Claude should not be called
-
-		// Verify the correct functions were called
-		expect(mockReadJSON).toHaveBeenCalledWith('test-tasks.json');
-		expect(mockWriteJSON).toHaveBeenCalled();
-		expect(mockGenerateTaskFiles).toHaveBeenCalled();
-
-		// Clean up
-		delete process.env.PERPLEXITY_API_KEY;
-	});
-
-	test('should append timestamp correctly in XML-like format', async () => {
-		// Mock streaming for successful response
-		const mockStream = {
-			[Symbol.asyncIterator]: jest.fn().mockImplementation(() => {
-				return {
-					next: jest
-						.fn()
-						.mockResolvedValueOnce({
-							done: false,
-							value: {
-								type: 'content_block_delta',
-								delta: {
-									text: 'Additional information about the subtask implementation.'
-								}
-							}
-						})
-						.mockResolvedValueOnce({ done: true })
-				};
-			})
-		};
-
-		mockCreate.mockResolvedValue(mockStream);
-
-		// Call the function
-		const result = await testUpdateSubtaskById(
-			'test-tasks.json',
-			'3.1',
-			'Add details about API endpoints'
-		);
-
-		// Verify the XML-like format with timestamp
-		expect(result).toBeDefined();
-		expect(result.details).toMatch(
-			/<info added on [0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}.[0-9]{3}Z>/
-		);
-		expect(result.details).toMatch(
-			/<\/info added on [0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}.[0-9]{3}Z>/
-		);
-
-		// Verify the same timestamp is used in both opening and closing tags
-		const openingMatch = result.details.match(
-			/<info added on ([0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}.[0-9]{3}Z)>/
-		);
-		const closingMatch = result.details.match(
-			/<\/info added on ([0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}.[0-9]{3}Z)>/
-		);
-
-		expect(openingMatch).toBeTruthy();
-		expect(closingMatch).toBeTruthy();
-		expect(openingMatch[1]).toBe(closingMatch[1]);
-	});
-
-	let mockTasksData;
-	const tasksPath = 'test-tasks.json';
-	const outputDir = 'test-tasks-output'; // Assuming generateTaskFiles needs this
-
-	beforeEach(() => {
-		// Reset mocks before each test
-		jest.clearAllMocks();
-
-		// Reset mock data (deep copy to avoid test interference)
-		mockTasksData = JSON.parse(
-			JSON.stringify({
-				tasks: [
-					{
-						id: 1,
-						title: 'Parent Task 1',
-						status: 'pending',
-						dependencies: [],
-						priority: 'medium',
-						description: 'Parent description',
-						details: 'Parent details',
-						testStrategy: 'Parent tests',
-						subtasks: [
-							{
-								id: 1,
-								title: 'Subtask 1.1',
-								description: 'Subtask 1.1 description',
-								details: 'Initial subtask details.',
-								status: 'pending',
-								dependencies: []
-							},
-							{
-								id: 2,
-								title: 'Subtask 1.2',
-								description: 'Subtask 1.2 description',
-								details: 'Initial subtask details for 1.2.',
-								status: 'done', // Completed subtask
-								dependencies: []
-							}
-						]
-					}
-				]
-			})
-		);
-
-		// Default mock behaviors
-		mockReadJSON.mockReturnValue(mockTasksData);
-		mockDirname.mockReturnValue(outputDir); // Mock path.dirname needed by generateTaskFiles
-		mockGenerateTaskFiles.mockResolvedValue(); // Assume generateTaskFiles succeeds
-	});
-
-	test('should successfully update subtask using Claude (non-research)', async () => {
-		const subtaskIdToUpdate = '1.1'; // Valid format
-		const updatePrompt = 'Add more technical details about API integration.'; // Non-empty prompt
-		const expectedClaudeResponse =
-			'Here are the API integration details you requested.';
-
-		// --- Arrange ---
-		// **Explicitly reset and configure mocks for this test**
-		jest.clearAllMocks(); // Ensure clean state
-
-		// Configure mocks used *before* readJSON
-		mockExistsSync.mockReturnValue(true); // Ensure file is found
-		mockGetAvailableAIModel.mockReturnValue({
-			// Ensure this returns the correct structure
-			type: 'claude',
-			client: { messages: { create: mockCreate } }
-		});
-
-		// Configure mocks used *after* readJSON (as before)
-		mockReadJSON.mockReturnValue(mockTasksData); // Ensure readJSON returns valid data
-		async function* createMockStream() {
-			yield {
-				type: 'content_block_delta',
-				delta: { text: expectedClaudeResponse.substring(0, 10) }
-			};
-			yield {
-				type: 'content_block_delta',
-				delta: { text: expectedClaudeResponse.substring(10) }
-			};
-			yield { type: 'message_stop' };
-		}
-		mockCreate.mockResolvedValue(createMockStream());
-		mockDirname.mockReturnValue(outputDir);
-		mockGenerateTaskFiles.mockResolvedValue();
-
-		// --- Act ---
-		const updatedSubtask = await taskManager.updateSubtaskById(
-			tasksPath,
-			subtaskIdToUpdate,
-			updatePrompt,
-			false
-		);
-
-		// --- Assert ---
-		// **Add an assertion right at the start to check if readJSON was called**
-		expect(mockReadJSON).toHaveBeenCalledWith(tasksPath); // <<< Let's see if this passes now
-
-		// ... (rest of the assertions as before) ...
-		expect(mockGetAvailableAIModel).toHaveBeenCalledWith({
-			claudeOverloaded: false,
-			requiresResearch: false
-		});
-		expect(mockCreate).toHaveBeenCalledTimes(1);
-		// ... etc ...
-	});
-
-	test('should successfully update subtask using Perplexity (research)', async () => {
-		const subtaskIdToUpdate = '1.1';
-		const updatePrompt = 'Research best practices for this subtask.';
-		const expectedPerplexityResponse =
-			'Based on research, here are the best practices...';
-		const perplexityModelName = 'mock-perplexity-model'; // Define a mock model name
-
-		// --- Arrange ---
-		// Mock environment variable for Perplexity model if needed by CONFIG/logic
-		process.env.PERPLEXITY_MODEL = perplexityModelName;
-
-		// Mock getAvailableAIModel to return Perplexity client when research is required
-		mockGetAvailableAIModel.mockReturnValue({
-			type: 'perplexity',
-			client: { chat: { completions: { create: mockChatCompletionsCreate } } } // Match the mocked structure
-		});
-
-		// Mock Perplexity's response
-		mockChatCompletionsCreate.mockResolvedValue({
-			choices: [{ message: { content: expectedPerplexityResponse } }]
-		});
-
-		// --- Act ---
-		const updatedSubtask = await taskManager.updateSubtaskById(
-			tasksPath,
-			subtaskIdToUpdate,
-			updatePrompt,
-			true
-		); // useResearch = true
-
-		// --- Assert ---
-		expect(mockReadJSON).toHaveBeenCalledWith(tasksPath);
-		// Verify getAvailableAIModel was called correctly for research
-		expect(mockGetAvailableAIModel).toHaveBeenCalledWith({
-			claudeOverloaded: false,
-			requiresResearch: true
-		});
-		expect(mockChatCompletionsCreate).toHaveBeenCalledTimes(1);
-
-		// Verify Perplexity API call parameters
-		expect(mockChatCompletionsCreate).toHaveBeenCalledWith(
-			expect.objectContaining({
-				model: perplexityModelName, // Check the correct model is used
-				temperature: 0.7, // From CONFIG mock
-				max_tokens: 4000, // From CONFIG mock
-				messages: expect.arrayContaining([
-					expect.objectContaining({
-						role: 'system',
-						content: expect.any(String)
-					}),
-					expect.objectContaining({
-						role: 'user',
-						content: expect.stringContaining(updatePrompt) // Check prompt is included
-					})
-				])
-			})
-		);
-
-		// Verify subtask data was updated
-		const writtenData = mockWriteJSON.mock.calls[0][1]; // Get data passed to writeJSON
-		const parentTask = writtenData.tasks.find((t) => t.id === 1);
-		const targetSubtask = parentTask.subtasks.find((st) => st.id === 1);
-
-		expect(targetSubtask.details).toContain(expectedPerplexityResponse);
-		expect(targetSubtask.details).toMatch(/<info added on .*>/); // Check for timestamp tag
-		expect(targetSubtask.description).toMatch(/\[Updated: .*]/); // Check description update
-
-		// Verify writeJSON and generateTaskFiles were called
-		expect(mockWriteJSON).toHaveBeenCalledWith(tasksPath, writtenData);
-		expect(mockGenerateTaskFiles).toHaveBeenCalledWith(tasksPath, outputDir);
-
-		// Verify the function returned the updated subtask
-		expect(updatedSubtask).toBeDefined();
-		expect(updatedSubtask.id).toBe(1);
-		expect(updatedSubtask.parentTaskId).toBe(1);
-		expect(updatedSubtask.details).toContain(expectedPerplexityResponse);
-
-		// Clean up env var if set
-		delete process.env.PERPLEXITY_MODEL;
-	});
-
-	test('should fall back to Perplexity if Claude is overloaded', async () => {
-		const subtaskIdToUpdate = '1.1';
-		const updatePrompt = 'Add details, trying Claude first.';
-		const expectedPerplexityResponse =
-			'Perplexity provided these details as fallback.';
-		const perplexityModelName = 'mock-perplexity-model-fallback';
-
-		// --- Arrange ---
-		// Mock environment variable for Perplexity model
-		process.env.PERPLEXITY_MODEL = perplexityModelName;
-
-		// Mock getAvailableAIModel: Return Claude first, then Perplexity
-		mockGetAvailableAIModel
-			.mockReturnValueOnce({
-				// First call: Return Claude
-				type: 'claude',
-				client: { messages: { create: mockCreate } }
-			})
-			.mockReturnValueOnce({
-				// Second call: Return Perplexity (after overload)
-				type: 'perplexity',
-				client: { chat: { completions: { create: mockChatCompletionsCreate } } }
-			});
-
-		// Mock Claude to throw an overload error
-		const overloadError = new Error('Claude API is overloaded.');
-		overloadError.type = 'overloaded_error'; // Match one of the specific checks
-		mockCreate.mockRejectedValue(overloadError); // Simulate Claude failing
-
-		// Mock Perplexity's successful response
-		mockChatCompletionsCreate.mockResolvedValue({
-			choices: [{ message: { content: expectedPerplexityResponse } }]
-		});
-
-		// --- Act ---
-		const updatedSubtask = await taskManager.updateSubtaskById(
-			tasksPath,
-			subtaskIdToUpdate,
-			updatePrompt,
-			false
-		); // Start with useResearch = false
-
-		// --- Assert ---
-		expect(mockReadJSON).toHaveBeenCalledWith(tasksPath);
-
-		// Verify getAvailableAIModel calls
-		expect(mockGetAvailableAIModel).toHaveBeenCalledTimes(2);
-		expect(mockGetAvailableAIModel).toHaveBeenNthCalledWith(1, {
-			claudeOverloaded: false,
-			requiresResearch: false
-		});
-		expect(mockGetAvailableAIModel).toHaveBeenNthCalledWith(2, {
-			claudeOverloaded: true,
-			requiresResearch: false
-		}); // claudeOverloaded should now be true
-
-		// Verify Claude was attempted and failed
-		expect(mockCreate).toHaveBeenCalledTimes(1);
-		// Verify Perplexity was called as fallback
-		expect(mockChatCompletionsCreate).toHaveBeenCalledTimes(1);
-
-		// Verify Perplexity API call parameters
-		expect(mockChatCompletionsCreate).toHaveBeenCalledWith(
-			expect.objectContaining({
-				model: perplexityModelName,
-				messages: expect.arrayContaining([
-					expect.objectContaining({
-						role: 'user',
-						content: expect.stringContaining(updatePrompt)
-					})
-				])
-			})
-		);
-
-		// Verify subtask data was updated with Perplexity's response
-		const writtenData = mockWriteJSON.mock.calls[0][1];
-		const parentTask = writtenData.tasks.find((t) => t.id === 1);
-		const targetSubtask = parentTask.subtasks.find((st) => st.id === 1);
-
-		expect(targetSubtask.details).toContain(expectedPerplexityResponse); // Should contain fallback response
-		expect(targetSubtask.details).toMatch(/<info added on .*>/);
-		expect(targetSubtask.description).toMatch(/\[Updated: .*]/);
-
-		// Verify writeJSON and generateTaskFiles were called
-		expect(mockWriteJSON).toHaveBeenCalledWith(tasksPath, writtenData);
-		expect(mockGenerateTaskFiles).toHaveBeenCalledWith(tasksPath, outputDir);
-
-		// Verify the function returned the updated subtask
-		expect(updatedSubtask).toBeDefined();
-		expect(updatedSubtask.details).toContain(expectedPerplexityResponse);
-
-		// Clean up env var if set
-		delete process.env.PERPLEXITY_MODEL;
-	});
-
-	// More tests will go here...
-});
+// End of file
